program that finds and prints the first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,7 +9,7 @@
 int main(void)
 {
 	int i;
-	unsigned long int fib, prv_fib1, prv_fib2;
+	unsigned int fib, prv_fib1, prv_fib2;
 
 	prv_fib1 = 0;
 	prv_fib2 = 1;
@@ -17,9 +17,9 @@ int main(void)
 	{
 		fib = prv_fib1 + prv_fib2;
 		if (i == 97)
-		printf("%lu", fib);
+		printf("%u", fib);
 		else
-		printf("%lu, ", fib);
+		printf("%u, ", fib);
 		prv_fib1 = prv_fib2;
 		prv_fib2 = fib;
 	}
